+++ container/build_in_container.sh
@@ -16,6 +16,8 @@ cat >> ./lennox_m30.idl <<'IDL'
 #pragma DCPS_DATA_TYPE "LxScheduleUpdateIDL::scheduleUpdate"
 #pragma DCPS_DATA_TYPE "LxManualAwayUpdateIDL::manualAwayUpdate"
 #pragma DCPS_DATA_TYPE "LxManualAwayStatusIDL::manualAwayStatus"
+#pragma DCPS_DATA_TYPE "LxAlertActiveIDL::alertActive"
+#pragma DCPS_DATA_TYPE "LxAlertClearedIDL::alertCleared"
 IDL
 
 "$ACE_ROOT/bin/mwc.pl" -type gnuace lennox_sub.mwc